Quizz_bot/requirements.txt
aiogram>=3.0.0
tortoise-orm>=0.19.3
aerich>=0.7.1
asyncpg>=0.28.0
python-dotenv>=1.0.0
